Add Vector class to represent translation vectors
A Vector Is Not A Point...

Vector class has x, y attributes, but these represent direction and
magnitude decomposed into cartesian coordinate components. (A Vector's
direction and magnitude being with respect to a cartesian 0,0 origin.)

Vectors can produce an affine transform that can be used to translate a
Point.

To accomodate this widening of the space for
things-that-can-be-used-to-transform, make some changes to the shared
examples for transformations that allow the interface tests to be
properly shared between objects.