Don't screw up -ve angles; Relax precision for 90º
There was a silly typo bug that meant that negative-angle arcs which
weren't one of the 90º increments went squiffy. Fix that.

Relax precision of 90º radians value to 12 d.p., which stops some angles
being considered > 90º when really it's just because of floating point
rounding errors down at 14 d.p. or so causing problems.

Also cull a dead method