Implement curve/line intersection checking
First cut implementation of curve/line intersection checking, based on
the JS bezier library from https://github.com/Pomax/bezierjs.

Separate intersection checking code out into subclasses by kind -
line/line and curve/line, add methods the CurveSegment and LineSegment
to report whether they're a curve or a line.

Curve/Line intersection is inherently a teeny bit fuzzy. Add classes for
approximation checking. Approximately (Approximately.equal?(1, 1.001))
and ApproximateRange (ApproximateRange.new(1..2).include?(2.0001))

There's a bug in the JS implementation this is based on which is
currently carried over here where, once aligned, if the y positions of the control
points are equal, and the y positions of the start and end points are
also equal there's a divide-by-zero error and everything goes to pot.

I'll address that in the next commit, but this is a big commit and I
don't want to layer any more complexity on…