Add presentation style object to pathlike things
And render the style object out to SVG. (Concentrating on fleshing this
out sufficiently before bringing PDF renderer back up to parity.)

Adding style means lots of API changes to Pathlike objects, which
affects a lot of files not hugely.

Also wound up removing the Curve class, because it had become an
inconvenient and extra-code-requiring convenvience.

Added Circle and CircleBuilder to give a more convenient way to create
circles with a specified centre, rather than having to create Arcs of
the full circle angle and having to figure out where to translate them
to get the centre point you wanted.