Rework path metadata; SVG parsing uses IDs.
Path metadata spread across several attributes was getting to be a real
pain to deal with safely when copying / altering pathlikes. This change
moves it all into a Metadata::Instance object and provides a mix in that
gives direct access. This simplifies Pathlike implementation so that it
only requires one method related to metadata, and adding new metadata
attributes will only require changes to the Metadata instance and its
mixin module.

This was all precipitated by needing to add an SVG `id` attribute analog,
and it was painful and irritating.

Improve SVG parser and renderer so that they support reading/writing the
SVG `id` attribute using the Metadata `name` attribute. This is
primarily to support using SVG fixtures for curve/line intersection
specs, so that the spec can be authored in a vector graphics editor like
Inkscape, Affinity Designer, or Illustrator.