intersection specs using authored SVG fixtures
Use the new SVG parser and specs authored in a vector graphics app to
write better intersection checker specs.

The best part of this is that prior to this, authoring meaningful
intersection checker specs - i.e. specs where the expected outcome can
be verified outside of doing the maths separately from the code - was
basically impossible, and i had very little confidence in testing this
code. Now I can eyeball the expected results in Affinity Designer or
Inkscape...