Add Curve/Curve intersection checking
This is basically just the iterative intersection finding from
Curve/Line, so there's plenty of scope for refactoring and DRYing up
this code.

The spec helpers for this got a go-over to make diffable output happen
and be useful, along with adding the ability to easily generate debug
output as an SVG showing the inputs and the expected and actual
intersections.

This was super useful in figuring out that the is-curve-flat threshold
needed bringing down a couple of orders of magnitude. (The code was
originally taken from a blog post about drawing vectors on screen, so
its definition of flat enough and mine turn out to be rather
different.)