Refactor intersection checking
Curve/Curve is only iterative, so the iterative intersection finder has
been extracted and is used for that and by the curve/line checker.

Add #line to LineSegment and CurveSegment for returning a LineSegment,
from start and end point of a curve. For a LineSegment it simply returns
self.

Add #split to CurveSegment, that passes through to DeCasteljau, so that
both line and curve have a split method. Yay polymorphism.

Remove the existing CurveCurve checker since it was just a shallow
rewrite of the iterative finder anyway.