Rearrange namespaces for Segments
Looking ahead to Path intersections, there's going to be classes
parallel to the Line/Curve segment intersection checkers, and the
Draught:: namespace is getting kind of crowded for these
similar-in-purpose-but-operating-on-different-objects classes.

So, move LineSegment and CurveSegment and builder friends under
Draught::Segment, and rename them (so, now it's Draught::Segment::Line
and Draught::Segment::Line::Builder, for example).

It feels both better and a bit icky (Segment::Line::Builder was
Segment::Builder::Line, and I'm still a little on the fence), so there
may be some more shuffling once I've been using them for a while.

Next step will be to move the Intersection Checker stuff under Segment
too, so there's a sensible place to put path intersection stuff and I
don't have to contort over the names.