Find intersections between polyline/spline paths
Add a graph of path segments for use in intersection finding, and the
PathIntersectionPoint class to represent a path intersection between two
paths, which points to the intersected paths as well as the intersection
point itself.

The intersection spec helper gets updated to handle this new kind of
intersection finding. It's a bit complicated now, and needs some
simplification (maybe splitting up so that the differences between
segment and path intersection don't have to be specially handled).

The path intersection finder itself uses a quadtree to cull potentially
intersectable segments down cheaply.